Make server logic vector clock-aware (#153)
- Modify server logic for queries and subscriptions to support vector
clock cursors
- Implement the changes from https://github.com/xmtp/proto/pull/206,
including splitting SID into two fields, using a dict for the vector
clock, and modifying node ID from uint16 to uint32

I was worried about type-conversions between int32 and uint32 and
overflow issues, but it looks like with two's complement, [overflowed
int32's behave the same as uint32's after
conversion](https://go.dev/play/p/pQmLPP56nzx) (or in other words, the
increment operation does exactly the same thing to the underlying byte
representation either way). In other words, our postgres schema can only
support signed int32's, but it seems like we can just pretend they are
uint32's and have the same effect.

We do still need to worry about the postgres sequence overflowing, as
well as the max value of uint32/uint64 - when implementing the
node-to-node syncing logic, my plan is to regularly log the node's
vector clock, so that we can see it coming from a long way away.

+++ contracts/src/Nodes.sol
@@ -14,9 +14,9 @@ All nodes on the network periodically check this contract to determine which nod
 contract Nodes is ERC721, Ownable {
     constructor() ERC721("XMTP Node Operator", "XMTP") Ownable(msg.sender) {}
 
-    // uint16 counter so that we cannot create more than 65k IDs
+    // uint32 counter so that we cannot create more than max IDs
     // The ERC721 standard expects the tokenID to be uint256 for standard methods unfortunately
-    uint16 private _nodeIdCounter;
+    uint32 private _nodeIdCounter;
 
     // A node, as stored in the internal mapping
     struct Node {
@@ -26,7 +26,7 @@ contract Nodes is ERC721, Ownable {
     }
 
     struct NodeWithId {
-        uint16 nodeId;
+        uint32 nodeId;
         Node node;
     }
 
@@ -42,8 +42,8 @@ contract Nodes is ERC721, Ownable {
         address to,
         bytes calldata signingKeyPub,
         string calldata httpAddress
-    ) public onlyOwner returns (uint16) {
-        uint16 nodeId = _nodeIdCounter;
+    ) public onlyOwner returns (uint32) {
+        uint32 nodeId = _nodeIdCounter;
         _mint(to, nodeId);
         _nodes[nodeId] = Node(signingKeyPub, httpAddress, true);
         _emitNodeUpdate(nodeId);
@@ -101,7 +101,7 @@ contract Nodes is ERC721, Ownable {
     Get a list of healthy nodes with their ID and metadata
      */
     function healthyNodes() public view returns (NodeWithId[] memory) {
-        uint16 totalNodeCount = _nodeIdCounter;
+        uint32 totalNodeCount = _nodeIdCounter;
         uint256 healthyCount = 0;
 
         // First, count the number of healthy nodes
@@ -116,7 +116,7 @@ contract Nodes is ERC721, Ownable {
         uint256 currentIndex = 0;
 
         // Populate the array with healthy nodes
-        for (uint16 i = 0; i < totalNodeCount; i++) {
+        for (uint32 i = 0; i < totalNodeCount; i++) {
             if (_nodeExists(i) && _nodes[i].isHealthy) {
                 healthyNodesList[currentIndex] = NodeWithId({
                     nodeId: i,
@@ -133,10 +133,10 @@ contract Nodes is ERC721, Ownable {
     Get all nodes regardless of their health status
      */
     function allNodes() public view returns (NodeWithId[] memory) {
-        uint16 totalNodeCount = _nodeIdCounter;
+        uint32 totalNodeCount = _nodeIdCounter;
         NodeWithId[] memory allNodesList = new NodeWithId[](totalNodeCount);
 
-        for (uint16 i = 0; i < totalNodeCount; i++) {
+        for (uint32 i = 0; i < totalNodeCount; i++) {
             allNodesList[i] = NodeWithId({nodeId: i, node: _nodes[i]});
         }
