+++ pkg/registry/interface.go
@@ -24,5 +24,5 @@ and notifying listeners when the list of nodes changes.
 type NodeRegistry interface {
 	GetNodes() ([]Node, error)
 	OnNewNodes() (<-chan []Node, CancelSubscription)
-	OnChangedNode(uint16) (<-chan Node, CancelSubscription)
+	OnChangedNode(uint32) (<-chan Node, CancelSubscription)
 }
